\name{obs}
\alias{obs}
\title{OBIS observations}
\usage{
obs(species = NULL, foptions = list())
}
\arguments{
  \item{species}{The species name for which we require
  data}

  \item{foptions}{Additional arugments we wil never use
  except for debugging}
}
\description{
A longer description
}
\examples{
\dontrun{
# obs(species = "Kogia breviceps")
}
}
\keyword{data}
\keyword{observations}

